function/requirements.txt
jsonpickle==1.3
aws-xray-sdk==2.4.3
pillow>=8.0.0
newspaper3k
boto3